requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.6
pydantic==1.10.7
aiofiles==23.1.0
aiofile==3.8.1
soundfile==0.12.1
librosa==0.10.0.post2
torchaudio==0.13.1
hydra-core>=1.2.0
einops==0.6.0
numpy==1.23.5
setuptools==67.7.2
wheel==0.40.0
nemo_toolkit[asr]==1.21.0
click>=8.0.0
h11==0.14.0
huggingface_hub==0.17.3
gTTS==2.3.1
pydub==0.25.1
streamlit
requests
